Percobaan 1

+++ Jobsheet7/SiakadFor24.java
@@ -1,10 +1,11 @@
-    package Jobsheet7;
+package Jobsheet7;
 
 public class SiakadFor24 {
 
     public static void main(String[] args) {
 
         java.util.Scanner sc = new java.util.Scanner(System.in);
+
         double nilai, tertinggi = 0, terendah = 100;
         for (int i = 1; i <= 10; i++) {
             System.out.print("Masukkan nilai ke-" + i + ": ");
@@ -16,8 +17,8 @@ public class SiakadFor24 {
                 terendah = nilai;
             }
         }
-        System.out.println("Nilai tertinggi adalah: " + (int) tertinggi);
-        System.out.println("Nilai terendah adalah: " + (int) terendah);
+        System.out.println("Nilai tertinggi adalah: " +  tertinggi);
+        System.out.println("Nilai terendah adalah: " +  terendah);
     }
 }
